GRIDEDIT-221: Move helper functions
