+++ src/MeshKernelApi/MeshKernel.cpp
@@ -51,6 +51,7 @@
 #include <MeshKernelApi/CurvilinearParameters.hpp>
 #include <MeshKernelApi/MeshKernel.hpp>
 #include <MeshKernelApi/SplinesToCurvilinearParameters.hpp>
+#include <MeshKernelApi/Utils.hpp>
 
 namespace meshkernelapi
 {
@@ -63,132 +64,6 @@ namespace meshkernelapi
 
     static char exceptionMessage[512] = "";
     static meshkernel::MeshGeometryError meshGeometryError = meshkernel::MeshGeometryError();
-
-    // TODO: Return result instead of relying on second input parameter
-    static void ConvertGeometryListToPointVector(const GeometryList& geometryListIn, std::vector<meshkernel::Point>& result)
-    {
-        if (geometryListIn.numberOfCoordinates == 0)
-        {
-            return;
-        }
-        result.resize(geometryListIn.numberOfCoordinates);
-
-        for (auto i = 0; i < geometryListIn.numberOfCoordinates; i++)
-        {
-            result[i] = {geometryListIn.xCoordinates[i], geometryListIn.yCoordinates[i]};
-        }
-    }
-
-    // TODO: Return result instead of relying on second input parameter
-    static void ConvertGeometryListToSampleVector(const GeometryList& geometryListIn, std::vector<meshkernel::Sample>& result)
-    {
-        if (geometryListIn.numberOfCoordinates == 0)
-        {
-            throw std::invalid_argument("MeshKernel: The samples are empty.");
-        }
-        result.resize(geometryListIn.numberOfCoordinates);
-
-        for (auto i = 0; i < geometryListIn.numberOfCoordinates; i++)
-        {
-            result[i] = {geometryListIn.xCoordinates[i], geometryListIn.yCoordinates[i], geometryListIn.zCoordinates[i]};
-        }
-    }
-
-    // TODO: Return result instead of relying on second input parameter
-    static void ConvertPointVectorToGeometryList(std::vector<meshkernel::Point> pointVector, GeometryList& result)
-    {
-        if (pointVector.size() < result.numberOfCoordinates)
-        {
-            throw std::invalid_argument("MeshKernel: Invalid memory allocation, the point-vector size is smaller than the number of coordinates.");
-        }
-
-        for (auto i = 0; i < result.numberOfCoordinates; i++)
-        {
-            result.xCoordinates[i] = pointVector[i].x;
-            result.yCoordinates[i] = pointVector[i].y;
-        }
-    }
-
-    static bool SetSplines(const GeometryList& geometryListIn, meshkernel::Splines& spline)
-    {
-        if (geometryListIn.numberOfCoordinates == 0)
-        {
-            return false;
-        }
-
-        std::vector<meshkernel::Point> splineCornerPoints;
-        ConvertGeometryListToPointVector(geometryListIn, splineCornerPoints);
-
-        const auto indices = FindIndices(splineCornerPoints, 0, splineCornerPoints.size(), meshkernel::doubleMissingValue);
-
-        for (const auto& index : indices)
-        {
-            const auto size = index[1] - index[0] + 1;
-            if (size > 0)
-            {
-                spline.AddSpline(splineCornerPoints, index[0], size);
-            }
-        }
-
-        return true;
-    }
-
-    static bool SetMeshGeometry(int meshKernelId, MeshGeometryDimensions& meshGeometryDimensions, MeshGeometry& meshGeometry)
-    {
-        if (meshKernelId >= meshInstances.size())
-        {
-            return false;
-        }
-
-        meshGeometry.nodex = &(meshInstances[meshKernelId]->m_nodex[0]);
-        meshGeometry.nodey = &(meshInstances[meshKernelId]->m_nodey[0]);
-        meshGeometry.nodez = &(meshInstances[meshKernelId]->m_nodez[0]);
-        meshGeometry.edge_nodes = &(meshInstances[meshKernelId]->m_edgeNodes[0]);
-
-        meshGeometryDimensions.maxnumfacenodes = meshkernel::maximumNumberOfNodesPerFace;
-        meshGeometryDimensions.numface = static_cast<int>(meshInstances[meshKernelId]->GetNumFaces());
-        if (meshGeometryDimensions.numface > 0)
-        {
-            meshGeometry.face_nodes = &(meshInstances[meshKernelId]->m_faceNodes[0]);
-            meshGeometry.facex = &(meshInstances[meshKernelId]->m_facesCircumcentersx[0]);
-            meshGeometry.facey = &(meshInstances[meshKernelId]->m_facesCircumcentersy[0]);
-            meshGeometry.facez = &(meshInstances[meshKernelId]->m_facesCircumcentersz[0]);
-        }
-
-        if (meshInstances[meshKernelId]->GetNumNodes() == 1)
-        {
-            meshGeometryDimensions.numnode = 0;
-            meshGeometryDimensions.numedge = 0;
-        }
-        else
-        {
-            meshGeometryDimensions.numnode = static_cast<int>(meshInstances[meshKernelId]->GetNumNodes());
-            meshGeometryDimensions.numedge = static_cast<int>(meshInstances[meshKernelId]->GetNumEdges());
-        }
-
-        return true;
-    }
-
-    static std::vector<meshkernel::Point> ComputeLocations(const MeshGeometryDimensions& meshGeometryDimensions, const MeshGeometry& meshGeometry, meshkernel::MeshLocations interpolationLocation)
-    {
-        std::vector<meshkernel::Point> locations;
-        if (interpolationLocation == meshkernel::MeshLocations::Nodes)
-        {
-            locations = meshkernel::ConvertToNodesVector(meshGeometryDimensions.numnode, meshGeometry.nodex, meshGeometry.nodey);
-        }
-        if (interpolationLocation == meshkernel::MeshLocations::Edges)
-        {
-            const auto edges = meshkernel::ConvertToEdgeNodesVector(meshGeometryDimensions.numedge, meshGeometry.edge_nodes);
-            const auto nodes = meshkernel::ConvertToNodesVector(meshGeometryDimensions.numnode, meshGeometry.nodex, meshGeometry.nodey);
-            locations = ComputeEdgeCenters(nodes, edges);
-        }
-        if (interpolationLocation == meshkernel::MeshLocations::Faces)
-        {
-            locations = meshkernel::ConvertToFaceCentersVector(meshGeometryDimensions.numface, meshGeometry.facex, meshGeometry.facey);
-        }
-
-        return locations;
-    }
 
     int HandleExceptions(const std::exception_ptr exceptionPtr)
     {
@@ -295,7 +170,7 @@ namespace meshkernelapi
 
             meshInstances[meshKernelId]->SetFlatCopies(meshkernel::Mesh2D::AdministrationOptions::AdministrateMeshEdges);
 
-            SetMeshGeometry(meshKernelId, meshGeometryDimensions, meshGeometry);
+            SetMeshGeometry(meshKernelId, meshGeometryDimensions, meshGeometry, meshInstances);
         }
         catch (...)
         {
@@ -315,7 +190,7 @@ namespace meshkernelapi
             }
             meshInstances[meshKernelId]->SetFlatCopies(meshkernel::Mesh2D::AdministrationOptions::AdministrateMeshEdgesAndFaces);
 
-            SetMeshGeometry(meshKernelId, meshGeometryDimensions, meshGeometry);
+            SetMeshGeometry(meshKernelId, meshGeometryDimensions, meshGeometry, meshInstances);
         }
         catch (...)
         {
